requirements.txt
matplotlib
pandas
pandas-ta
requests
schedule
notion-client
openai
tweepy
python-dotenv
feedparser
